Added information about installing "elliptic" and "crypto-js" Node.js libraries
The "elliptic" and "crypto-js" Node.js libraries are needed to do Public/Private Key Encryption for the Addresses and Signature signing/verification. Research on doing this may be seen in the "Nodes/research/FranklinCryptoUtility.js" file along with the referenced hyperlinks explaining where the information was obtained. Read the "Nodes/research/FranklinCryptoUtility.js" file on how to import and install the "elliptic" and "crypto-js" Node.js libraries.