Updated Code to set Current Dificulty when /peers/notify-new-block received
Updated code to set the Current Difficulty of the Node to the Peer Node from which a /peers/notify-new-block RESTFul Web Service call was received. Found out I had to do this, because you could have situation where two Nodes are connected before any Mining is done on any of thye Nodes and if the Current Difficulty is NOT set on the receiving Node, then a Miner connected to this Node will not be Mining with the correct Current Difficulty.