Fixed code to remove pending transactions placed in blockchain
Fixed bug in code whereby the Pending Transactions of a Node were not being removed after synchronizing with Peer's Chain and removing from the Pending List those Transactions that were placed in a Block. Also, when miner submits a mined block, added code to check if any of the transactions in the Mined Block from the Miner havce already been placed inside of a Block in the Blockchain.